Don't fail uninstall on missing dirs

+++ bin/setup/system-roms-retroarch.sh
@@ -100,18 +100,24 @@ uninstall() {
   # Remove core options
   while IFS="$tab" read library_name; do
     local retroarch_emulator_config_dir="$retroarch_config_dir/config/$library_name"
-    find "$retroarch_emulator_config_dir" -name '*.opt' -exec rm -rf "{}" \;
+    if [ -d "$retroarch_emulator_config_dir" ]; then
+      find "$retroarch_emulator_config_dir" -name '*.opt' -exec rm -rf "{}" \;
+    fi
   done < <(system_setting 'select(.emulators) | .emulators[] | select(.library_name) | .library_name')
 
   # Remove retroarch remappings
   local remapping_dir=$(crudini --get "$retropie_system_config_dir/retroarch.cfg" '' 'input_remapping_directory' 2>/dev/null || true)
   if [ -n "$remapping_dir" ]; then
     remapping_dir=${remapping_dir//\"/}
-    find "$remapping_dir" -name '*.rmp' -exec rm -rf "{}" \;
+    if [ -d "$remapping_dir" ]; then
+      find "$remapping_dir" -name '*.rmp' -exec rm -rf "{}" \;
+    fi
   fi
 
   # Remove retroarch configs
-  find "$HOME/RetroPie/roms/$system/" -name '*.cfg' -exec rm -rf "{}" \;
+  if [ -d "$HOME/RetroPie/roms/$system/" ]; then
+    find "$HOME/RetroPie/roms/$system/" -name '*.cfg' -exec rm -rf "{}" \;
+  fi
 }
 
 "$1" "${@:3}"
